requirements_dev.txt
numpy
xarray
# Testing
pytest
gitpython
